refactor(frontend): move i18n resources
