+++ frontend/public/locales/en/index.json
@@ -5,8 +5,6 @@
   "personal-info-desc": "In this section, you can see and update your personal information.",
   "upload": "Upload document for CDCP",
   "upload-desc": "In this section, you can your upload documents.",
-  "intake": "Intake",
-  "intake-desc": "In this section, you can apply for your CDCP application.",
   "view-letters": "My letters",
   "view-letters-desc": "View letters about your CDCP application.",
   "view-cdcp": "View my CDCP messages",
