+++ frontend/public/locales/fr/apply-adult-child.json
@@ -242,6 +242,7 @@
     "contact-info": "Coordonnées",
     "comm-prefs": "Communication",
     "dental-insurance": "Accès à d'autres assurances ou couvertures de soins dentaires",
+    "dental-benefit-has-access": "Accès à des prestations dentaires gouvernementales\u00a0:",
     "dob": "Date de naissance",
     "sin": "Numéro d'assurance sociale (NAS)",
     "full-name": "Nom complet",
