+++ frontend/public/locales/fr/renew-adult-child.json
@@ -120,7 +120,7 @@
     "address-field": {
       "address": "Adresse postale",
       "mailing-address": "Adresse postale",
-      "apartment": "Inclure le numéro d'appartement (s'il y a lieu), le numéro municipal et le nom de la rue. Par exemple : 4B - 123 Rue Main",
+      "address-note": "Inclure le numéro d'appartement (s'il y a lieu), le numéro municipal et le nom de la rue. Par exemple\u00a0: 4B - 123 Rue Main",
       "country": "Pays",
       "province": "Province, territoire, État ou région",
       "city": "Ville ou municipalité",
@@ -173,8 +173,7 @@
       "save-btn": "Sauvegarder"
     },
     "error-message": {
-      "federal-benefit-required": "Sélectionnez si vous avez la prestation dentaire fédérale ",
-      "provincial-benefit-required": "Sélectionnez si vous avez la prestation dentaire provinciale ou territoriale"
+      "federal-provincial-territorial-benefit-required": "Sélectionnez si vous bénéficiez de prestations dentaires fédérales provinciale ou territoriale"
     }
   },
   "update-dental-benefits": {
